Added custom time filter Organizing files Broke out ticket types into their own file Started stubbing in new data formats
