+++ package.json
@@ -19,7 +19,6 @@
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
     "@types/styled-components": "^5.1.26",
-    "framer-motion": "^11.14.4",
     "localtunnel": "^2.0.2",
     "qr-scanner": "^1.4.2",
     "react": "^18.2.0",
